[secure_boot] Verify signer at build time and in verify-firmware CLI
Unified follow-up that supersedes #494 and #495 — both addressed the same
underlying problem (verifying *who* signed a Secure Boot v2 image, not
just "somebody"), but split across build-time and CLI surfaces with
duplicated key files and incompatibly reshaped APIs. This collapses them
to a single source of truth.

# What it does

1. **Catches wrong-env firmware bundled into local desktop builds** (#494's fix).
   `frostsnapp/rust/build.rs` now verifies the bytes at
   `target/.../{env}-frontier.bin` are signed by the env's expected
   secure-boot key. Reproduces the v0.3.0 incident as a build-time panic
   instead of a silent broken `.app`.

2. **Categorises `verify-firmware` output** (#495's fix). The CLI now
   distinguishes:
     - `✅ Verified — signed by Frostsnap PROD key` (exit 0)
     - `✅ Verified — signed by Frostsnap DEV key (not production!)` (exit 0)
     - `⚠️  SIGNED BY UNKNOWN KEY` (exit 1)
     - `⚠️  NOT SIGNED` / `⚠️  SIGNATURE INVALID / CORRUPT` (exit 1)
   Malformed firmware bodies are surfaced best-effort so a corrupt body
   can't hide the signer verdict.

# Single source of truth per env

`frostsnap_factory/bootloader/{env}/secure-boot-key.pem` is the one file
build.rs and the CLI both consult. Dev's was already there (private key);
prod's is added by this commit (SubjectPublicKeyInfo public key — the
matching private key stays offline). Same filename per env, content type
is a parsing detail. `secure_boot_pubkey_from_pem` accepts either form.

This drops:
- #494's separate `bootloader/prod/secure-boot-public-key.pem` filename.
- #495's parallel `secure_boot_key/prod/public_key.pem` directory.
- #495's committed `key_digest.hex` cache files (digests are derived at
  startup via `compute_key_digest`, so the PEM is the only authority).

The .gitignore line that blocked the path is removed — the file is now
legitimately committable, accidental private-key commits get caught at
review.

# Crate extraction

`secure_boot.rs` moves out of `frostsnap_factory` (which depends on
mysql, clap, tracing-subscriber — none of which belong in a build script)
into a new `frostsnap_secure_boot` crate. Build.rs can now consume the
verifier without dragging the factory tool's deps into the desktop build.
`sign_firmware` takes its RNG by argument so the crate stays `rand`-free
(only `rand_core`); factory's CLI passes `rand::thread_rng()` at the call
site.

# API surface

- `verify_firmware(bytes) -> Result<Verified, VerifyError>` where
  `Verified { public_key, key_digest, image_digest }`. One entry point,
  carries everything both callers need.
- `Signer { Prod, Dev, Unknown }` + `classify_signer(key_digest, prod, dev)`.
  Pinned digests passed in rather than read from disk so this crate
  imposes no filesystem layout on its callers.
- `compute_key_digest(&RsaPublicKey)` — derives the ESP32 eFuse key
  digest from a pubkey using the same `block[36..812]` layout the
  verifier reads, so digests can be pinned from PEMs at startup.
- `secure_boot_pubkey_from_pem(pem)` — accepts either a PKCS1/PKCS8
  private key (dev) or a SubjectPublicKeyInfo public key (prod), returns
  `RsaPublicKey` either way.
- `VerifyError::is_not_signed()` distinguishes "not a Secure Boot image
  at all" (TooSmall / NotSectorAligned / BadMagic) from "signed but
  corrupt." Drives the CLI's NOT SIGNED vs CORRUPT split.

# Test plan

- [x] `cargo test -p frostsnap_secure_boot` — 7 pass (sign/verify
      aligned and unaligned, PEM loader accepts both forms, compute_key_digest
      matches verify output, classify_signer routing, two espsecure
      fixture tests).
- [x] `cargo test -p frostsnap_factory` — passes.
- [x] `cargo run -p frostsnap_factory -- verify-firmware --input
      target/.../dev-frontier.bin` → `✅ Verified — signed by Frostsnap DEV key`.
- [x] `cargo run -p frostsnap_factory -- verify-firmware --input
      frostsnap_factory/bootloader/prod/signed-bootloader.bin` →
      `✅ Verified — signed by Frostsnap PROD key`. This proves the
      committed `bootloader/prod/secure-boot-key.pem` produces the same
      key digest as the prod-signed bootloader's embedded signing key —
      i.e. the new PEM matches the actual prod key.
- [x] Build.rs end-to-end matrix verified locally in #494's testing,
      same logic preserved here.

Closes #494, closes #495.
