+++ frostsnap_factory/Cargo.toml
@@ -12,8 +12,8 @@ rand.workspace = true
 aes = "0.8"
 cbc = "0.1"
 hmac = "0.12"
-crc.workspace = true
 sha2.workspace = true
+frostsnap_secure_boot.workspace = true
 frostsnap_coordinator.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
