+++ frostsnap_factory/bootloader/prod/secure-boot-key.pem
@@ -1,0 +1,11 @@
+-----BEGIN PUBLIC KEY-----
+MIIBojANBgkqhkiG9w0BAQEFAAOCAY8AMIIBigKCAYEAsZuG0fC2wVTqZELlOInf
+1nx1wzKpTHbnjedgRKR9V9UBYKXVeY2j+DSUnvp3Y8HnJkQTt2SSX3cgMGJIK7SS
+NaGcZEkxJHGwuhYegEzacnKvVgnT8kOjL5a/+OIQDhZ/91nGgtfH9EJtU4PrDN/T
+ZC5CMvTkfnCgERMyTIkjQtlL15AAo1kBgUh8BTHCEg+xFVS7MTLpE9p1DipZe5+H
+IDysm06pfXn0/r3KS9wMlMFMvJb6igwYBrNA+S+cf89Ac7rtNRrq8LtVm2v8GxNf
+EZyCo03QAlkg35BYaea/lF+bFjcO+MXWunrwzw67+ey6BgxtNsupPljszdigay1N
+M2xK8Fa2Meg+EXsi44HwDeftkTAwv0uLG99Qn3RwcXah0oU9ieUuOiajTi12l23j
+fNMOAReMVU1s/t3rmgDAY0PT5fcgtSjTDbVUWVh70yPvr7rnVnmwBUs2h1SOKrAR
+o+p4WqjCOcQv9s7qdZMgRzcCD8+VAgXu5l5cB6AVGWVPAgMBAAE=
+-----END PUBLIC KEY-----
