+++ frostsnap_secure_boot/tests/fixtures/test-firmware-signed.bin
