+++ frostsnap_secure_boot/tests/fixtures/test-firmware.bin
