+++ frostsnapp/rust/Cargo.toml
@@ -25,6 +25,9 @@ time = { version = "0.3.36", features = ["formatting"] }
 bip21 = {version = "0.5.0", default-features = false}
 futures.workspace = true
 
+[build-dependencies]
+frostsnap_secure_boot.workspace = true
+
 [dependencies.flutter_rust_bridge]
 version = "=2.11.1"
 default-features = false
